requirements.txt
# Core data handling
pandas>=2.0
numpy>=1.24
openpyxl>=3.1           # required by pandas.read_excel for .xlsx inputs

# NLP preprocessing
spacy>=3.7,<3.8
nltk>=3.8
unidecode>=1.3
# The Spanish model is not installed from requirements.txt — it is fetched
# via `python -m spacy download es_core_news_md` in reproduce.sh so that
# pip resolves a version compatible with the installed spaCy.

# LLM classification (Azure OpenAI)
openai>=1.10,<2.0
python-dotenv>=1.0      # optional .env loading in cam_peru.llm_client

# Embeddings & clustering
transformers>=4.35,<5.0
torch>=2.0
umap-learn>=0.5.5
hdbscan>=0.8.33
scikit-learn>=1.3
scipy>=1.10

# Network analysis
networkx>=3.1

# Visualization
matplotlib>=3.7
seaborn>=0.12
wordcloud>=1.9

# Translation + inter-rater agreement (notebooks)
deep-translator>=1.11
krippendorff>=0.6
